split advanced and status codes into separate examples
